Improve backward compatibility with previous plugin versions
